create a for loop to add all cards in array
using a for loop (see syntax below) an index variable called i
is used to loop over the indices in the constant array.

For each one of the indices we call the function addSpecificCard
which grabs the element from the array and inserts it into the
DOM just as before, using the "innerHTML" property of the continer
node.

https://storage.googleapis.com/quest_editor_uploads/67LhBYqnydJJZxnbNgd6Og9M9RYC6UcG.png